It boots!
In QEMU, at least. Haven't tried it on real hardware yet.

This took several days and tons of scattered tutorials, but I now have
a working OS for the Raspberry Pi. The goal is to have an OS written
in some Forth dialect, but there are two major challenges before I can
start on that:

1. Keyboard input. This will require USB support, which is going to be
   tricky (to put it mildly).

2. Memory management. Right now, everything is running on physical
   addresses, but to be a proper OS, I'll need to activate the MMU.
   I'm not sure how that's going to work, but I think that perhaps
   I should figure that out *before* I start in on the USB driver.