The MMU works!
After being stuck on enabling the MMU for over a week, I ended up
patching QEMU to track down the small inconsistency with how I was
interpreting flags. Turns out that level 3 desciptors invert the
"this is a page table" bit, and that block memory expects to have
that bit set for that level. I'm going to refactor the'
`PageTableDescriptor` wrapper to handle this better, but I want to
ensure these changes are kept around so I have a known good state
to restore to if (when) I mess it up again.